Clean Up and Format

+++ src/main/java/com/playonlinux/core/observer/ObservableDirectoryFiles.java
@@ -38,7 +38,6 @@ public class ObservableDirectoryFiles extends ObservableDirectory<File[]> {
         observableDirectoryThread = new ObservableDirectoryThread(this);
     }
 
-
     @Override
     public void shutdown() {
         observableDirectoryThread.stopChecking();
@@ -50,13 +49,12 @@ public class ObservableDirectoryFiles extends ObservableDirectory<File[]> {
         this.observableDirectoryThread.start();
     }
 
-
     /* TODO: Test this method with better coverage */
     protected File[] findFiles() {
         File[] files = observedDirectory.listFiles();
         List<File> filesFiltered = new LinkedList<>();
         assert files != null;
-        for(File file: files) {
+        for (File file : files) {
             if (!file.getName().startsWith(".")) {
                 filesFiltered.add(file);
             }
@@ -64,7 +62,6 @@ public class ObservableDirectoryFiles extends ObservableDirectory<File[]> {
 
         return filesFiltered.toArray(new File[filesFiltered.size()]);
     }
-
 
     private class ObservableDirectoryThread extends Thread {
         private final ObservableDirectoryFiles observableDirectoryFiles;
@@ -89,14 +86,14 @@ public class ObservableDirectoryFiles extends ObservableDirectory<File[]> {
         @Override
         public void run() {
             File[] lastDirectoryContent = null;
-            while(this.isRunning()) {
-                if(observedDirectory.exists()) {
+            while (this.isRunning()) {
+                if (observedDirectory.exists()) {
                     File[] directoryContent = observableDirectoryFiles.findFiles();
                     if (!Arrays.equals(directoryContent, lastDirectoryContent)) {
                         this.observableDirectoryFiles.notifyObservers(directoryContent);
                     }
                     try {
-                        Thread.sleep((long) checkInterval);
+                        Thread.sleep(checkInterval);
                     } catch (InterruptedException e) {
                         LOGGER.debug(e);
                         this.stopChecking();
